Dialogue almost fully implemented
Can now control all aspects of dialogue and the player's triggers. Animation and sount can be setup but won't currently do anything in the scene. There are a few glitches to iron out, but I'm aware of most of them. too add more triggers for the player, open the playerTriggersDatabase and add more. If you add more triggers, the script probably won't work anymore but I know how to fix it.
Search for two commented lines in the dialogue editor script that have the .Clear command and un-comment them. save and click on the dialogue npc in Unity. go back and re-comment the lines, save and return to unity. I will fix this, just don't have time right now